requirements.txt
numpy~=1.23.1
jax~=0.3.17
flax~=0.6.0
netket~=3.5.1
matplotlib~=3.5.2
scipy~=1.9.2
PyYAML~=6.0
tqdm~=4.64.1